dispensable状态时, 速度测试变得严格

+++ turquoise.imageloader/src/main/java/sviolet/turquoise/x/imageloader/entity/ServerSettings.java
@@ -637,7 +637,7 @@ public class ServerSettings implements ComponentManager.Component{
         return values.abortOnLowNetworkSpeedBoundaryProgress;
     }
 
-    public float getAbortOnLowNetworkSpeedDeadline(){
+    public long getAbortOnLowNetworkSpeedDeadline(){
         return values.abortOnLowNetworkSpeedDeadline;
     }
